backend/requirements.txt
apscheduler>=3.10.4
pymysql>=1.1.0
matplotlib>=3.8.0
numpy>=1.26.0
plotly>=5.22.0
kaleido>=0.2.1
cairosvg>=2.7.0
requests>=2.32.0
duckdb>=1.0.0

fastapi>=0.115.0
uvicorn[standard]>=0.27.0
clerk-backend-api>=1.2.2
python-dotenv>=1.0.1

pandas>=2.1.0
antlr4-python3-runtime>=4.13.1